+++ source/Messaging.Infrastructure/OutgoingMessages/Common/Xml/HeaderWriter.cs
@@ -15,9 +15,10 @@
 using System;
 using System.Threading.Tasks;
 using System.Xml;
+using Messaging.Application.Common;
 using Messaging.Domain.OutgoingMessages;
 
-namespace Messaging.Application.Common;
+namespace Messaging.Infrastructure.OutgoingMessages.Common.Xml;
 
 public static class HeaderWriter
 {
