+++ source/Messaging.Infrastructure/OutgoingMessages/GenericNotification/GenericNotificationDocumentWriter.cs
@@ -18,6 +18,7 @@ using System.Threading.Tasks;
 using System.Xml;
 using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages.GenericNotification;
+using Messaging.Infrastructure.OutgoingMessages.Common.Xml;
 
 namespace Messaging.Infrastructure.OutgoingMessages.GenericNotification;
