+++ source/Messaging.Tests/Application/OutgoingMessages/DocumentFactoryTests.cs
@@ -16,6 +16,7 @@ using System.Threading.Tasks;
 using Messaging.Application.Common;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Domain.OutgoingMessages;
+using Messaging.Infrastructure.OutgoingMessages.Common.Xml;
 using NodaTime;
 using Xunit;
